// src/Model/Types/Grades/DetailedUEResult.ts
/** Resultats detaillés d'une UE */
export type DetailedUEResult = {
    /** Moyenne de l'UE */
    UEResult: number;
    /** Moyenne global du pôle CC de l'UE */
    CCResult: number;
    /** Moyenne global du pôle SAE de l'UE */
    SAEResult: number;
    /** Moyenne de chaque ressources du pôle CC de l'UE */
    AllCCResults: number[];
    /** Moyenne de chaque ressources du pôle SAE de l'UE */
    AllSAEResults: number[];
};